requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
openai==1.59.7
pydantic==2.10.5
pydantic-settings==2.7.1
python-dotenv==1.0.1
groq==0.16.0
google-generativeai==0.8.4
PyMuPDF==1.25.3
Pillow==11.1.0
sse-starlette==2.2.1
python-multipart==0.0.20
